requirements-dev.txt
django == 5.2.4
pillow == 11.3.0
cryptography == 45.0.5
djangorestframework == 3.16.0
pillow == 11.3.0
drf-yasg == 1.21.10
djangorestframework-simplejwt == 5.5.1